requirements.txt
PySide6>=6.5.0
markdown2>=2.4.0
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
requests>=2.28.0
pyinstaller>=5.13.0